refactor: useChartVoteHandler 커스텀 훅 생성
