+++ src/pages/List/components/DonationsModal/DonationsModal.jsx
@@ -42,7 +42,7 @@ function DonationsModal({
     closeModal
   );
 
-  useEscapeModal();
+  useEscapeModal(closeModal);
 
   if (!isOpen) return null;
